Better sharing / separation of free, pro code
The classes folder was generally common to both PressPermit and PressPermit Pro, but contained a number of differences.  This is now corrected, with all distinct code contained in the following folders and files:

* /modules contains Pro files
* /vendor is currently only used for Pro
* /includes contains PressPermit files
* /includes-pro contains PressPermit Pro files
* presspermit.php
* presspermit-pro.php
* readme.txt is different for each package

Note that /languages already contains translations for shared code only (/classes folder), while each Pro module has its own language folder and file.
