+++ classes/PublishPress/Permissions/UI/Settings.php
@@ -24,6 +24,11 @@ class Settings
         wp_enqueue_script('presspermit-settings', PRESSPERMIT_URLPATH . "/common/js/settings{$suffix}.js", ['jquery', 'jquery-form'], PRESSPERMIT_VERSION, true);
         $wp_scripts->in_footer[] = 'presspermit-settings';  // otherwise it will not be printed in footer  @todo: review
 
+        if (defined('PRESSPERMIT_PRO_VERSION')) {
+            wp_enqueue_script('presspermit-pro-settings', PRESSPERMIT_URLPATH . "/includes-pro/settings-pro{$suffix}.js", ['jquery', 'jquery-form'], PRESSPERMIT_VERSION, true);
+            $wp_scripts->in_footer[] = 'presspermit-pro-settings';  // otherwise it will not be printed in footer  @todo: review
+        }
+
         if (!current_user_can('pp_manage_settings'))
             wp_die(PWP::__wp('Cheatin&#8217; uh?'));
